refactor(st-clock): code size optimization
Re-ordering structures to avoid gaps and minimize data.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/323783
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>

+++ drivers/st/clk/clk-stm32mp13.c
@@ -1694,7 +1694,7 @@ static const struct stm32_clk_pll *clk_st32_pll_data(unsigned int idx)
 }
 
 struct stm32_pll_cfg {
-	int pll_id;
+	uint8_t pll_id;
 };
 
 static unsigned long clk_stm32_pll_recalc_rate(struct stm32_clk_priv *priv,  int id,
@@ -1780,8 +1780,8 @@ static const struct stm32_clk_ops clk_stm32_pll_ops = {
 }
 
 struct clk_stm32_composite_cfg {
-	int gate_id;
-	int div_id;
+	uint8_t gate_id;
+	uint8_t div_id;
 };
 
 static unsigned long clk_stm32_composite_recalc_rate(struct stm32_clk_priv *priv,
